requirements.txt
fastapi
uvicorn
jinja2
python-multipart
youtube-transcript-api
google-generativeai
python-dotenv
requests
